Improve widget pref migration
