+++ mobile/src/main/java/org/openhab/habdroid/ui/PreferencesActivity.kt
@@ -395,12 +395,7 @@ class PreferencesActivity : AbstractBaseActivity() {
                 iconFormatPreference.setOnPreferenceChangeListener { pref, _ ->
                     val context = pref.context
                     clearImageCache(context)
-                    val ids = AppWidgetManager.getInstance(context)
-                        .getAppWidgetIds(ComponentName(context, ItemUpdateWidget::class.java))
-                    val intent = Intent(context, ItemUpdateWidget::class.java)
-                        .setAction(AppWidgetManager.ACTION_APPWIDGET_UPDATE)
-                        .putExtra(AppWidgetManager.EXTRA_APPWIDGET_IDS, ids)
-                    context.sendBroadcast(intent)
+                    ItemUpdateWidget.updateAllWidgets(context)
                     true
                 }
             }
